[v1.0] WebShop Application Scenario
This commit contains the sources of the WebShop and Retailer Application
which is
running on SAP HANA Cloud
https://cloudmodelespmhana.hana.ondemand.com/espm-cloud-web/webshop

The corresponding WebShop Scenario description on SCN
Find the link of new ESPM Java WebShop (+ Retailer, + ESPM Cloud Model)
scenario tutorial here
http://scn.sap.com/community/developer-center/cross-technology

This tutorial explains what the ESPM Model is and how it is build in
Java as basis for the WebShop application.
Furthermore by means of the so caller Retailer applications it is
explained how Roles and Security concepts of HANA Cloud Web applications
are explained.